refactor: rename and move fs to node-fs
All direct use of node-fs outside of node-server must be refactored
away in preparation for web extension support.
