+++ src/server/server.ts
@@ -16,11 +16,11 @@ import { doHover } from "./features/hover/hover";
 import { provideReferences } from "./features/references";
 import { doSignatureHelp } from "./features/signature-help/signature-help";
 import { searchWorkspaceSymbol } from "./features/workspace-symbols/workspace-symbol";
+import { findFiles } from "./node-fs";
 import ScannerService from "./scanner";
 import type { ISettings } from "./settings";
 import StorageService from "./storage";
 import { getSCSSRegionsDocument } from "./utils/embedded";
-import { findFiles } from "./utils/fs";
 
 interface InitializationOption {
 	workspace: string;
