+++ src/test/features/diagnostics.spec.ts
@@ -3,8 +3,8 @@ import fs, { Stats } from "fs";
 import { stub, SinonStub } from "sinon";
 import { DiagnosticSeverity, DiagnosticTag } from "vscode-languageserver-types";
 import { doDiagnostics } from "../../server/features/diagnostics/diagnostics";
+import * as fsUtils from "../../server/node-fs";
 import StorageService from "../../server/storage";
-import * as fsUtils from "../../server/utils/fs";
 import { EXTENSION_NAME } from "../../shared/constants";
 import * as helpers from "../helpers";
